requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
pytz==2024.1
pandas==2.2.1